Moved PGDSelect into its own file.
The PGDSelect class was getting big and hard to maintain inside the
main script so I put it in its own file.  The AA3to1 check was disabled
as it no longer worked without easy access to that variable.
